test(SBTL): move SBTL Catch2 tests into src/Tests/SBTL-Tests.cpp
ibell asked for a dedicated test file so rebases against master don't
keep hitting CoolProp-Tests.cpp (a high-churn shared file).  All 10
SBTL test cases (the [SBTL] tag family plus the hidden
[.][hermite_accuracy] benchmark) move to src/Tests/SBTL-Tests.cpp.

Mechanical move:
  - SBTL-Tests.cpp gets its own ENABLE_CATCH guard, Catch2 include,
    and the SBTL-relevant headers (AbstractState, DataStructures,
    SBTLBackend) plus <algorithm>/<cmath>/<random>/<vector>/etc.
  - CoolProp-Tests.cpp loses 496 lines (the SBTL block) and the
    <random> include that only [hermite_accuracy] needed.
  - CMakeLists.txt: list SBTL-Tests.cpp the same way CoolProp-Tests.cpp
    is listed — REMOVE from APP_SOURCES at the top (so it's not pulled
    into the main library), then re-add in the COOLPROP_TEST and
    COOLPROP_CATCH_MODULE blocks where the test binaries are built.

[SBTL] suite from the new file: 9/9 pass, 726/726 assertions green.

Pre-commit hook regex updated locally to gate on the new file too.
