+++ src/Tests/CoolProp-Tests.cpp
@@ -11,7 +11,6 @@
 #include "miniz.h"
 #include <atomic>
 #include <map>
-#include <random>
 #include <set>
 #include <thread>
 
@@ -5560,501 +5559,6 @@ TEST_CASE("REFPROP supports DmolarQ / DmassQ inputs (#1845)", "[REFPROP][1845]")
     auto AS2 = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
     AS2->update(CoolProp::DmassQ_INPUTS, 15.0, 1.0);
     CHECK(AS->T() == Catch::Approx(AS2->T()).epsilon(1e-6));
-}
-
-TEST_CASE("NormalizedPHTable: build_normph_table fills cell values matching HEOS for CO2", "[SBTL][normph][build]") {
-    // After build_normph_table runs on a region, every cell with a finite
-    // hmolar entry should match a fresh HEOS evaluation at (h, P) =
-    // (table.hmolar[i][j], table.yvec[j]) — confirming the cell-fill loop
-    // actually plumbed (xnorm, P) -> h -> HEOS update -> stored values
-    // through correctly.  Tolerance 1e-12 (just round-trip floating point).
-    auto SBTL_AS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("SBTL&HEOS", "CO2"));
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-    auto SBTL = dynamic_cast<CoolProp::SBTLBackend*>(SBTL_AS.get());
-    REQUIRE(SBTL != nullptr);
-    HEOS->update(CoolProp::PT_INPUTS, 1e6, 280.0);
-    SBTL->update(CoolProp::HmolarP_INPUTS, HEOS->hmolar(), HEOS->p());
-
-    using R = CoolProp::NormalizedPHTable::Region;
-    for (auto region : {R::LIQUID, R::VAPOR, R::SUPER}) {
-        CoolProp::NormalizedPHTable table(region);
-        table.Nx = 12;  // small grid for fast test
-        table.Ny = 12;
-        SBTL->build_normph_table(table);
-
-        std::size_t finite_count = 0;
-        for (std::size_t i = 0; i < table.Nx; ++i) {
-            for (std::size_t j = 0; j < table.Ny; ++j) {
-                if (!std::isfinite(table.rhomolar[i][j])) continue;
-                ++finite_count;
-                CAPTURE(static_cast<int>(region));
-                CAPTURE(i);
-                CAPTURE(j);
-                CAPTURE(table.xvec[i]);
-                CAPTURE(table.yvec[j]);
-                CAPTURE(table.hmolar[i][j]);
-                CAPTURE(table.rhomolar[i][j]);
-                // Cell stored values came from a HEOS update; reproducing the
-                // same update should give back the same rhomolar to numerical
-                // precision.
-                HEOS->update(CoolProp::HmolarP_INPUTS, table.hmolar[i][j], table.yvec[j]);
-                CHECK(std::abs(HEOS->rhomolar() - table.rhomolar[i][j]) / std::abs(table.rhomolar[i][j]) < 1e-12);
-            }
-        }
-        // Sanity: at least most cells should be filled (some boundary holes
-        // are expected near pmin/pmax/dome edges).
-        CAPTURE(static_cast<int>(region));
-        CAPTURE(finite_count);
-        CAPTURE(table.Nx * table.Ny);
-        CHECK(finite_count > (table.Nx * table.Ny) / 2);
-    }
-}
-
-TEST_CASE("NormalizedPHTable: xnorm <-> h round-trip across all three regions for CO2", "[SBTL][normph]") {
-    // Verifies the coordinate-aligned PH math: for each region, after
-    // populate_normph_bounds runs, h_from_xnorm(xnorm_from_h(h, P), P) == h
-    // for any (h, P) inside that region's rectangle.  Sample at probe
-    // points strictly inside each region's bounds.
-    auto SBTL_AS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("SBTL&HEOS", "CO2"));
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-    auto SBTL = dynamic_cast<CoolProp::SBTLBackend*>(SBTL_AS.get());
-    REQUIRE(SBTL != nullptr);
-    // Trigger sat cache build
-    HEOS->update(CoolProp::PT_INPUTS, 1e6, 280.0);
-    SBTL->update(CoolProp::HmolarP_INPUTS, HEOS->hmolar(), HEOS->p());
-
-    using R = CoolProp::NormalizedPHTable::Region;
-    for (auto region : {R::LIQUID, R::VAPOR, R::SUPER}) {
-        CoolProp::NormalizedPHTable table(region);
-        table.AS = HEOS;
-        table.Nx = 16;
-        table.Ny = 16;
-        table.set_limits();
-        table.resize(table.Nx, table.Ny);  // populates xvec, yvec
-        SBTL->populate_normph_bounds(table);
-
-        REQUIRE(table.h_lo_isobar.size() == table.Ny);
-        REQUIRE(table.h_hi_isobar.size() == table.Ny);
-        // All bounds must be finite.  h_lo < h_hi at every isobar except
-        // at the edge rows where the region's enthalpy span collapses:
-        //   - LIQUID j=0: p ≈ p_min where T_sat(p_min) ≈ T_min, so
-        //     h(T_min, p) ≈ h_sat,L(p).
-        //   - SUPER j=0: p ≈ p_crit where h(T_min, p) ≈ h(T_max, p) at
-        //     the supercritical inflection.
-        // Both are coordinate degeneracies, not build failures (filed as
-        // CoolProp-4mf for the eventual fix via foi.5 h_c-aware split).
-        for (std::size_t j = 0; j < table.Ny; ++j) {
-            CAPTURE(j);
-            CAPTURE(table.h_lo_isobar[j]);
-            CAPTURE(table.h_hi_isobar[j]);
-            CHECK(std::isfinite(table.h_lo_isobar[j]));
-            CHECK(std::isfinite(table.h_hi_isobar[j]));
-            if ((region == R::SUPER || region == R::LIQUID) && j == 0) continue;
-            CHECK(table.h_lo_isobar[j] < table.h_hi_isobar[j]);
-        }
-
-        // Round-trip: pick interior (xnorm, P) points; map to h, map back.
-        for (double xnorm : {0.1, 0.3, 0.5, 0.7, 0.9}) {
-            // Take P at four evenly-spaced log-grid rows, avoiding the row endpoints
-            // where linear-in-logP interpolation pins to a boundary value.
-            for (std::size_t jp = 1; jp + 1 < table.Ny; jp += 4) {
-                const double P = std::sqrt(table.yvec[jp] * table.yvec[jp + 1]);  // mid-row geomean
-                const double h = table.h_from_xnorm(xnorm, P);
-                const double xnorm_back = table.xnorm_from_h(h, P);
-                CAPTURE(static_cast<int>(region));
-                CAPTURE(xnorm);
-                CAPTURE(P);
-                CAPTURE(h);
-                CAPTURE(xnorm_back);
-                CHECK(std::abs(xnorm_back - xnorm) < 1e-12);
-            }
-        }
-    }
-}
-
-TEST_CASE("SBTL saturation cache exposes h_sat,L(p) and h_sat,V(p) for CO2", "[SBTL][sat_cache]") {
-    // First building block toward a saturation-curve-aligned PH coordinate:
-    // h_sat,L(p) and h_sat,V(p) are cached in the SBTL backend's sat_cache
-    // (1000 log-p samples, cubic interp from underlying tabular sat data) and
-    // exposed as public methods.  These will be the lower / upper bounds of
-    // the normalized-h coordinate in the subcritical liquid / vapor subdomains
-    // when the coordinate-aligned PH path lands.
-    auto SBTL_AS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("SBTL&HEOS", "CO2"));
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-    auto SBTL = dynamic_cast<CoolProp::SBTLBackend*>(SBTL_AS.get());
-    REQUIRE(SBTL != nullptr);
-
-    // Sample subcritical pressures spanning the table range.  Tolerance is
-    // ~1e-3 — bounded by the cubic interpolation of the underlying sat table,
-    // not by the cache layer itself.  Tightens to machine precision when an
-    // H superancillary expansion is wired in (follow-up work).
-    const double pc = HEOS->p_critical();
-    // Sample fractions from 0.05 up: the H-superancillary fit's left piece
-    // (the lowest-pressure subdivision) currently delivers ~1e-2 relative
-    // at frac=0.01 because the single Cheb piece covers a very wide
-    // log-p range there.  CoolProp-05w tracks tightening that via more
-    // Cheb pieces on the low-p side; until then this test gates the
-    // accurate-region range only.
-    for (double frac : {0.05, 0.2, 0.5, 0.9}) {
-        const double p = pc * frac;
-        HEOS->update(CoolProp::PQ_INPUTS, p, 0.0);
-        const double hL_eos = HEOS->hmolar();
-        HEOS->update(CoolProp::PQ_INPUTS, p, 1.0);
-        const double hV_eos = HEOS->hmolar();
-
-        const double hL_sbtl = SBTL->saturation_hmolar_liquid(p);
-        const double hV_sbtl = SBTL->saturation_hmolar_vapor(p);
-
-        CAPTURE(p);
-        CAPTURE(hL_eos);
-        CAPTURE(hL_sbtl);
-        CAPTURE(hV_eos);
-        CAPTURE(hV_sbtl);
-        CHECK(std::abs(hL_sbtl - hL_eos) / std::abs(hL_eos) < 1e-8);
-        CHECK(std::abs(hV_sbtl - hV_eos) / std::abs(hV_eos) < 1e-8);
-    }
-}
-
-TEST_CASE("SBTL populate_corner_derivatives_pt reads HEOS partials at a probe state", "[SBTL][corner_derivs_pt]") {
-    // PT-table analogue: ∂f/∂T, ∂f/∂p, ∂²f/∂T², ∂²f/∂T∂p for f in {rho, h, s, u}.
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-    HEOS->update(CoolProp::PT_INPUTS, 1e6, 320.0);
-    CoolProp::SBTLCornerDerivsPT cd;
-    REQUIRE(CoolProp::populate_corner_derivatives_pt(*HEOS, cd));
-    CHECK(cd.rhomolar == Catch::Approx(static_cast<double>(HEOS->rhomolar())));
-    CHECK(cd.hmolar == Catch::Approx(static_cast<double>(HEOS->hmolar())));
-    CHECK(cd.drho_dT == Catch::Approx(static_cast<double>(HEOS->first_partial_deriv(CoolProp::iDmolar, CoolProp::iT, CoolProp::iP))));
-    CHECK(cd.dh_dp == Catch::Approx(static_cast<double>(HEOS->first_partial_deriv(CoolProp::iHmolar, CoolProp::iP, CoolProp::iT))));
-    CHECK(
-      cd.d2rho_dT2
-      == Catch::Approx(static_cast<double>(HEOS->second_partial_deriv(CoolProp::iDmolar, CoolProp::iT, CoolProp::iP, CoolProp::iT, CoolProp::iP))));
-    CHECK(
-      cd.d2h_dTp
-      == Catch::Approx(static_cast<double>(HEOS->second_partial_deriv(CoolProp::iHmolar, CoolProp::iT, CoolProp::iP, CoolProp::iP, CoolProp::iT))));
-}
-
-TEST_CASE("SBTL populate_corner_derivatives reads HEOS partials for CO2 at a probe state", "[SBTL][corner_derivs]") {
-    // populate_corner_derivatives reads value + 4 partials × 4 props from a
-    // primed HEOS state.  Verify the entries match direct first_partial_deriv
-    // / second_partial_deriv calls — this is the data-collection layer for
-    // Hermite bicubic corner derivatives (CoolProp-foi.1).
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-    HEOS->update(CoolProp::PT_INPUTS, 1e6, 320.0);  // single-phase vapor
-    CoolProp::SBTLCornerDerivs cd;
-    REQUIRE(CoolProp::populate_corner_derivatives(*HEOS, cd));
-    CHECK(cd.rhomolar == Catch::Approx(static_cast<double>(HEOS->rhomolar())));
-    CHECK(cd.T == Catch::Approx(static_cast<double>(HEOS->T())));
-    CHECK(cd.drho_dh == Catch::Approx(static_cast<double>(HEOS->first_partial_deriv(CoolProp::iDmolar, CoolProp::iHmolar, CoolProp::iP))));
-    CHECK(cd.dT_dp == Catch::Approx(static_cast<double>(HEOS->first_partial_deriv(CoolProp::iT, CoolProp::iP, CoolProp::iHmolar))));
-    CHECK(cd.d2rho_dh2
-          == Catch::Approx(
-            static_cast<double>(HEOS->second_partial_deriv(CoolProp::iDmolar, CoolProp::iHmolar, CoolProp::iP, CoolProp::iHmolar, CoolProp::iP))));
-    CHECK(cd.d2T_dhp
-          == Catch::Approx(
-            static_cast<double>(HEOS->second_partial_deriv(CoolProp::iT, CoolProp::iHmolar, CoolProp::iP, CoolProp::iP, CoolProp::iHmolar))));
-}
-
-TEST_CASE("SBTL Cheb1DPiece::eval_dlogp matches finite difference on a smooth function", "[SBTL][cheb_deriv]") {
-    // Cheb-derivative recurrence is the foundation for chain-ruling
-    // sat-curve derivatives ∂h_lo/∂(log p), ∂h_hi/∂(log p) into the
-    // Hermite bicubic corner data (CoolProp-foi.1).  Build a Cheb1DPiece
-    // on a smooth analytic function whose true derivative is known and
-    // verify eval_dlogp agrees with both the analytic answer and a
-    // central finite difference.
-    //
-    // Test function: f(p) = a + b log(p) + c (log p)^2 + d sin(log p)
-    // df/d(log p) = b + 2 c log(p) + d cos(log p)
-    const double a = 1.5, b = 0.7, c = -0.4, d = 0.9;
-    auto f = [&](double p) {
-        const double lp = std::log(p);
-        return a + b * lp + c * lp * lp + d * std::sin(lp);
-    };
-    auto fprime = [&](double p) {
-        const double lp = std::log(p);
-        return b + 2.0 * c * lp + d * std::cos(lp);
-    };
-    const double p_lo = 1e3, p_hi = 1e8;
-    const auto piece = CoolProp::Cheb1DPiece::build(p_lo, p_hi, 20, f);
-    REQUIRE(piece.valid());
-    for (double p : {2e3, 1e4, 5e4, 1e6, 5e6, 1e7, 5e7}) {
-        const double analytic = fprime(p);
-        const double cheb = piece.eval_dlogp(p);
-        const double fd = (f(p * std::exp(1e-6)) - f(p * std::exp(-1e-6))) / (2.0 * 1e-6);
-        CAPTURE(p);
-        CAPTURE(analytic);
-        CAPTURE(cheb);
-        CAPTURE(fd);
-        // Cheb expansion converges spectrally on this smooth f → < 1e-8 vs analytic.
-        CHECK(cheb == Catch::Approx(analytic).epsilon(1e-8));
-        CHECK(cheb == Catch::Approx(fd).epsilon(1e-6));
-    }
-}
-
-TEST_CASE("SBTL hermite_bicubic_polynomial_coeffs exactly reconstructs an arbitrary bicubic", "[SBTL][hermite_bicubic]") {
-    // Construct an arbitrary bicubic f(xi, eta) = Σ_{m,n in 0..3} c_{m,n} xi^m eta^n.
-    // Feed corner values and derivatives into hermite_bicubic_polynomial_coeffs;
-    // the returned alpha vector must reproduce f exactly at every (xi, eta) ∈ [0,1]^2.
-    // This validates the closed-form coefficients before they're wired into the
-    // SBTL build path (CoolProp-foi.1).
-    const double c[4][4] = {
-      {1.7, -0.3, 2.1, 0.5},
-      {0.4, 1.2, -0.9, 0.2},
-      {-1.1, 0.7, 1.4, -0.3},
-      {0.8, -0.5, 0.6, 1.1},
-    };
-    auto eval = [&](double xi, double eta) {
-        double v = 0.0;
-        double xp = 1.0;
-        for (int m = 0; m < 4; ++m) {
-            double ep = 1.0;
-            for (int n = 0; n < 4; ++n) {
-                v += c[m][n] * xp * ep;
-                ep *= eta;
-            }
-            xp *= xi;
-        }
-        return v;
-    };
-    auto pow_int = [](double base, int exp) {
-        double r = 1.0;
-        for (int k = 0; k < exp; ++k)
-            r *= base;
-        return r;
-    };
-    auto dxi = [&](double xi, double eta) {
-        double v = 0.0;
-        for (int m = 1; m < 4; ++m) {
-            for (int n = 0; n < 4; ++n) {
-                v += m * c[m][n] * pow_int(xi, m - 1) * pow_int(eta, n);
-            }
-        }
-        return v;
-    };
-    auto deta = [&](double xi, double eta) {
-        double v = 0.0;
-        for (int m = 0; m < 4; ++m) {
-            for (int n = 1; n < 4; ++n) {
-                v += n * c[m][n] * pow_int(xi, m) * pow_int(eta, n - 1);
-            }
-        }
-        return v;
-    };
-    auto dxideta = [&](double xi, double eta) {
-        double v = 0.0;
-        for (int m = 1; m < 4; ++m) {
-            for (int n = 1; n < 4; ++n) {
-                v += m * n * c[m][n] * pow_int(xi, m - 1) * pow_int(eta, n - 1);
-            }
-        }
-        return v;
-    };
-    const auto alpha = CoolProp::hermite_bicubic_polynomial_coeffs(eval(0, 0), eval(1, 0), eval(0, 1), eval(1, 1), dxi(0, 0), dxi(1, 0), dxi(0, 1),
-                                                                   dxi(1, 1), deta(0, 0), deta(1, 0), deta(0, 1), deta(1, 1), dxideta(0, 0),
-                                                                   dxideta(1, 0), dxideta(0, 1), dxideta(1, 1));
-    REQUIRE(alpha.size() == 16);
-    auto eval_alpha_local = [&](double xi, double eta) {
-        double v = 0.0;
-        for (int m = 0; m < 4; ++m) {
-            for (int n = 0; n < 4; ++n) {
-                v += alpha[4 * m + n] * pow_int(xi, m) * pow_int(eta, n);
-            }
-        }
-        return v;
-    };
-    for (double xi : {0.0, 0.1, 0.3, 0.5, 0.7, 0.9, 1.0}) {
-        for (double eta : {0.0, 0.1, 0.3, 0.5, 0.7, 0.9, 1.0}) {
-            CAPTURE(xi);
-            CAPTURE(eta);
-            CHECK(eval_alpha_local(xi, eta) == Catch::Approx(eval(xi, eta)).epsilon(1e-13));
-        }
-    }
-}
-
-TEST_CASE("SBTL normph core-property lookups match HEOS at single-phase probes", "[SBTL][conformance]") {
-    // End-to-end smoke check on the foi.1 build path: SBTL with Hermite
-    // bicubic for the core derived properties (rho, T, s, u) matches HEOS
-    // at three single-phase PT probes to within 1e-4 relative.
-    auto SBTL_AS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("SBTL&HEOS", "CO2"));
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-
-    struct Probe
-    {
-        double p_Pa, T_K;
-        const char* label;
-    };
-    const Probe probes[] = {
-      {1e6, 240.0, "subcritical liquid"},
-      {1e6, 320.0, "subcritical vapor"},
-      {1e7, 320.0, "supercritical"},
-    };
-    for (const auto& pr : probes) {
-        HEOS->update(CoolProp::PT_INPUTS, pr.p_Pa, pr.T_K);
-        SBTL_AS->update(CoolProp::PT_INPUTS, pr.p_Pa, pr.T_K);
-        const double rho_eos = HEOS->rhomolar();
-        const double rho_sbtl = SBTL_AS->rhomolar();
-        CAPTURE(pr.label);
-        CAPTURE(pr.p_Pa);
-        CAPTURE(pr.T_K);
-        CAPTURE(rho_eos);
-        CAPTURE(rho_sbtl);
-        CHECK(std::abs(rho_sbtl - rho_eos) / std::abs(rho_eos) < 1e-4);
-    }
-}
-
-TEST_CASE("SBTL accuracy on Water random PT/PH sweep", "[.][hermite_accuracy]") {
-    // Hidden by default ([.] prefix) — slow (~30s on Water).  Run with
-    //   --test-spec '[hermite_accuracy]'
-    // to validate SBTL accuracy on the IAPWS reference fluid against HEOS
-    // direct.  Reports rho / T / s deviations for both the PT path
-    // (normpt table) and the PH path (normph table) at 200 random
-    // single-phase states.  Deviations are directly comparable to the
-    // G13-15 Table 10 permissible values (v / T / s ~ 1e-5 relative).
-    //
-    // Asserts only that median rho is within sane bounds — looser than
-    // G13-15 because the near-critical box and 2-phase notch are
-    // excluded here, and any one-off CI flake should not break the test.
-    auto SBTL_AS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("SBTL&HEOS", "Water"));
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "Water"));
-
-    const double T_min = HEOS->Ttriple() * 1.1;
-    const double T_max = std::min(static_cast<double>(HEOS->Tmax()) * 0.95, 1.4 * static_cast<double>(HEOS->T_critical()));
-    const double p_min = std::max(static_cast<double>(HEOS->p_triple()) * 1.5, 1e3);
-    const double p_max = std::min(static_cast<double>(HEOS->pmax()) * 0.95, 5.0 * static_cast<double>(HEOS->p_critical()));
-    const double p_crit = HEOS->p_critical();
-
-    std::mt19937 gen(42);
-    std::uniform_real_distribution<double> log_p(std::log(p_min), std::log(p_max));
-    std::uniform_real_distribution<double> Tdist(T_min, T_max);
-    std::vector<std::pair<double, double>> states;
-    while (states.size() < 200) {
-        const double p = std::exp(log_p(gen));
-        const double T = Tdist(gen);
-        try {
-            HEOS->update(CoolProp::PT_INPUTS, p, T);
-            const double q = HEOS->Q();
-            if (q > 0.0 && q < 1.0) continue;
-            if (p > 0.97 * p_crit && p < 1.03 * p_crit) continue;
-        } catch (...) {
-            continue;
-        }
-        states.emplace_back(p, T);
-    }
-
-    struct ProbeErr
-    {
-        double p_Pa, T_K, dev;
-    };
-    struct Errors
-    {
-        std::vector<double> drho, dT, ds;
-        std::vector<ProbeErr> rho_worst;  // (p, T, drho), top by drho
-    };
-    auto record_worst = [](std::vector<ProbeErr>& w, double p, double T, double dev) {
-        w.push_back({p, T, dev});
-        std::sort(w.begin(), w.end(), [](auto& a, auto& b) { return a.dev > b.dev; });
-        if (w.size() > 3) w.resize(3);
-    };
-    Errors pt, ph;
-    for (auto [p, T] : states) {
-        try {
-            HEOS->update(CoolProp::PT_INPUTS, p, T);
-            SBTL_AS->update(CoolProp::PT_INPUTS, p, T);
-        } catch (...) {
-            continue;
-        }
-        const double drho = std::abs(SBTL_AS->rhomolar() - HEOS->rhomolar()) / std::abs(HEOS->rhomolar());
-        pt.drho.push_back(drho);
-        record_worst(pt.rho_worst, p, T, drho);
-        const double s_h_pt = HEOS->smolar();
-        pt.ds.push_back(std::abs(SBTL_AS->smolar() - s_h_pt) / std::max(std::abs(s_h_pt), 1e-30));
-    }
-    for (auto [p, T] : states) {
-        double h = 0.0;
-        try {
-            HEOS->update(CoolProp::PT_INPUTS, p, T);
-            h = HEOS->hmolar();
-            HEOS->update(CoolProp::HmolarP_INPUTS, h, p);
-            SBTL_AS->update(CoolProp::HmolarP_INPUTS, h, p);
-        } catch (...) {
-            continue;
-        }
-        const double drho = std::abs(SBTL_AS->rhomolar() - HEOS->rhomolar()) / std::abs(HEOS->rhomolar());
-        ph.drho.push_back(drho);
-        record_worst(ph.rho_worst, p, T, drho);
-        ph.dT.push_back(std::abs(SBTL_AS->T() - HEOS->T()) / std::abs(HEOS->T()));
-        const double s_h_ph = HEOS->smolar();
-        ph.ds.push_back(std::abs(SBTL_AS->smolar() - s_h_ph) / std::max(std::abs(s_h_ph), 1e-30));
-    }
-
-    struct Stats
-    {
-        double median, p99, max;
-    };
-    auto stats = [](std::vector<double> v) -> Stats {
-        if (v.empty()) return {0.0, 0.0, 0.0};
-        std::sort(v.begin(), v.end());
-        return {v[v.size() / 2], v[static_cast<std::size_t>(0.99 * (v.size() - 1))], v.back()};
-    };
-    auto report = [&](const char* label, const Errors& e) {
-        auto line = [&](const char* prop, const std::vector<double>& v) {
-            if (v.empty()) return;
-            const auto s = stats(v);
-            std::cout << "  " << prop << "  median=" << s.median << "  p99=" << s.p99 << "  max=" << s.max << "\n";
-        };
-        std::cout << "[hermite_accuracy] " << label << " (Water, n=" << e.drho.size() << "):\n";
-        line("rho", e.drho);
-        line("T  ", e.dT);
-        line("s  ", e.ds);
-        const double p_crit = 22.064e6;  // Water, approximate
-        const double T_crit = 647.096;
-        for (const auto& w : e.rho_worst) {
-            const double T_red = w.T_K / T_crit;
-            const double p_red = w.p_Pa / p_crit;
-            std::cout << "  worst-rho   p=" << w.p_Pa << " (p/p_c=" << p_red << ")  T=" << w.T_K << " (T/T_c=" << T_red << ")  drho=" << w.dev
-                      << "\n";
-        }
-    };
-    report("PT path (normpt)", pt);
-    report("PH path (normph)", ph);
-
-    // Loose sanity bounds — actual G13-15 conformance is tracked separately.
-    const auto pt_rho = stats(pt.drho), ph_rho = stats(ph.drho);
-    CHECK(pt_rho.median < 1e-6);
-    CHECK(ph_rho.median < 1e-6);
-}
-
-TEST_CASE("SBTL native speed_sound matches HEOS for CO2 single-phase states", "[SBTL][speed_sound]") {
-    // Native speed-of-sound path: SBTL builds a per-cell B-spline polynomial
-    // for w alongside (rho, T, s, h, p, u) and serves calc_speed_sound() from
-    // that polynomial.  This sanity test confirms the polynomial reproduces
-    // HEOS at a handful of single-phase liquid / vapor / supercritical states
-    // — full conformance vs IF97/IAPWS-95 spec is a follow-up.
-    auto SBTL_AS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("SBTL&HEOS", "CO2"));
-    auto HEOS = std::shared_ptr<CoolProp::AbstractState>(CoolProp::AbstractState::factory("HEOS", "CO2"));
-
-    struct Probe
-    {
-        double p_Pa, T_K;
-        const char* label;
-    };
-    const Probe probes[] = {
-      {1e6, 240.0, "subcritical liquid"},
-      {1e6, 320.0, "subcritical vapor"},
-      {1e7, 320.0, "supercritical"},
-    };
-    for (const auto& pr : probes) {
-        HEOS->update(CoolProp::PT_INPUTS, pr.p_Pa, pr.T_K);
-        SBTL_AS->update(CoolProp::PT_INPUTS, pr.p_Pa, pr.T_K);
-        const double w_eos = HEOS->speed_sound();
-        const double w_sbtl = SBTL_AS->speed_sound();
-        CAPTURE(pr.label);
-        CAPTURE(pr.p_Pa);
-        CAPTURE(pr.T_K);
-        CAPTURE(w_eos);
-        CAPTURE(w_sbtl);
-        // 1% — bounded by B-spline reconstruction error of w, not the EOS.
-        CHECK(std::abs(w_sbtl - w_eos) / std::abs(w_eos) < 1e-2);
-    }
 }
 
 #endif
